# apps/api/src/openapi.yaml
openapi: 3.1.0
info:
  title: Peter Dashboard API
  description: |
    Backend API for Peter dashboard platform.
    Handles dashboard management, BigQuery query execution, authentication, and lineage tracking.
  version: 0.1.0
  contact:
    name: Peter Team
    email: team@peter.com
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api-staging.peter.com
    description: Staging environment
  - url: https://api.peter.com
    description: Production environment
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: peter_session
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    DateTime:
      type: string
      format: date-time
      example: "2025-10-30T10:15:30Z"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid dashboard YAML"
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string
          example: "abc123"
        remediation:
          type: string
          example: "Check YAML syntax and try again"

    # Auth models
...
    ViewType:
      type: string
      enum: [analytical, operational, strategic]
      example: "analytical"
    DashboardMetadata:
      type: object
      required:
        - id
        - slug
        - name
        - owner_id
        - version
        - view_type
        - created_at
        - updated_at
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
          example: "revenue-dashboard"
        name:
          type: string
          example: "Revenue Dashboard"
        description:
          type: string
          nullable: true
          example: "Monthly revenue trends by region"
        owner_id:
          $ref: "#/components/schemas/UUID"
        owner_email:
          type: string
          format: email
          example: "user@example.com"
        storage_path:
          type: string
          example: "dashboards/revenue-dashboard/v1.yaml"
        version:
          type: integer
          minimum: 1
          example: 1
        view_type:
          $ref: "#/components/schemas/ViewType"
        tags:
          type: array
          items:
            type: string
          example: ["finance", "kpi"]
        created_at:
          $ref: "#/components/schemas/DateTime"
        updated_at:
          $ref: "#/components/schemas/DateTime"
        last_accessed:
          $ref: "#/components/schemas/DateTime"
          nullable: true
        access_count:
          type: integer
          example: 42
    DashboardYAML:
      type: string
      description: "Raw YAML content of dashboard definition"
      example: |
        version: 0
        kind: dashboard
        slug: revenue-dashboard
        title: "Revenue Dashboard"
        layout:
          - id: rev_kpi
            type: kpi
            query_ref: q_revenue
        queries:
          - id: q_revenue
            sql: "SELECT SUM(revenue) FROM mart.revenue_daily"
    DashboardCreate:
      type: object
      required:
        - yaml_content
      properties:
        yaml_content:
          $ref: "#/components/schemas/DashboardYAML"
        commit_message:
          type: string
          example: "Initial dashboard creation"
    DashboardUpdate:
      type: object
      required:
        - yaml_content
      properties:
        yaml_content:
          $ref: "#/components/schemas/DashboardYAML"
        commit_message:
          type: string
          example: "Updated chart colors"
    DashboardListResponse:
      type: object
      required:
        - dashboards
        - total
      properties:
        dashboards:
          type: array
          items:
            $ref: "#/components/schemas/DashboardMetadata"
        total:
          type: integer
          example: 10
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20

    # Validation & Compilation
    ValidationResult:
      type: object
      required:
        - valid
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: object
            properties:
              line:
                type: integer
              column:
                type: integer
              message:
                type: string
              severity:
                type: string
                enum: [error, warning]
    ExecutionPlan:
      type: object
      required:
        - queries
        - lineage_seeds
      properties:
        queries:
          type: array
          items:
            type: object
            properties:
              query_id:
                type: string
              sql:
                type: string
              depends_on:
                type: array
                items:
                  type: string
        lineage_seeds:
          type: array
          items:
            type: object
            properties:
              query_id:
                type: string
              tables:
                type: array
                items:
                  type: string

    # SQL Execution
...
paths:
...
  /api/v1/dashboards:
    get:
      tags: [dashboards]
      summary: List dashboards
      operationId: listDashboards
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
        - name: owner_id
          in: query
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: List of dashboards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardListResponse"
    post:
      tags: [dashboards]
      summary: Create dashboard
      operationId: createDashboard
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardCreate"
      responses:
        "201":
          description: Dashboard created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetadata"
        "400":
          description: Invalid YAML
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/dashboards/validate:
    post:
      tags: [dashboards]
      summary: Validate dashboard YAML
      operationId: validateDashboard
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              $ref: "#/components/schemas/DashboardYAML"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"
  /api/v1/dashboards/compile:
    post:
      tags: [dashboards]
      summary: Compile dashboard to execution plan
      operationId: compileDashboard
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              $ref: "#/components/schemas/DashboardYAML"
      responses:
        "200":
          description: Execution plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionPlan"
  /api/v1/dashboards/{slug}:
    get:
      tags: [dashboards]
      summary: Get dashboard
      operationId: getDashboard
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetadata"
        "404":
          description: Dashboard not found
    put:
      tags: [dashboards]
      summary: Update dashboard
      operationId: updateDashboard
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardUpdate"
      responses:
        "200":
          description: Dashboard updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetadata"
    delete:
      tags: [dashboards]
      summary: Delete dashboard
      operationId: deleteDashboard
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Dashboard deleted

  # SQL endpoints
...